+++ code/engine.vc2008/xrPhysics/PHStaticGeomShell.h
@@ -20,7 +20,7 @@ virtual		void			EnableObject		(CPHObject* obj){CPHUpdateObject::Activate();}
 virtual		dGeomID			dSpacedGeom			(){return dSpacedGeometry();}
 virtual		void			PhDataUpdate		(dReal step);
 virtual		void			PhTune				(dReal step){}
-virtual		void			InitContact			(dContact* c,bool& do_collide,u16 /*material_idx_1*/,u16 /*material_idx_2*/){}
+virtual		void			InitContact			(dContact* c,bool& do_collide, size_t /*material_idx_1*/, size_t /*material_idx_2*/){}
 virtual		u16				get_elements_number				()								{return 0;};
 virtual		CPHSynchronize	*get_element_sync				(u16 element)					{return NULL;};		
 public:
